Rename to squad_data and add  a base import having all the connections to the premier league website.
