+++ lib/tasks/import.rake
@@ -1,11 +1,11 @@
-require_relative '../scrapper/get_table_ranking'
+require_relative '../scrapper/squad_data'
 
 namespace :scrapper do
   desc "Initializing The Scrapper"
 
   task run_scrapper: :environment do
     scrapper = Scrapper::Scrape.new
-    p scrapper.db_import_team
+    p scrapper.fetch_table_data
     # Team.import(t)
   end
 end
